Pow no longer returns 1 when 0⁰ is given.
pow now panics in case of 0⁰
check_pow now returns None in case of 0⁰

pow and checked_pow now requires the trait 'Zero' as we need it to
check if base is zero when exp is zero.

Added basic seperate tests to check that basic behaivor works, and
that bought pow and checked_pow behaves correclty when 0⁰ is given.

Updated the documentation to state behavior in 0⁰ case.

Fixes #77