wrapping: remove some WrappingNeg tests
Some WrappingNeg tests rely on core::ops::Neg being implemented for
core::num::Wrapping<_>. Since it was only added in Rust 1.10, this
causes the build to fail for version 1.8 (the MSRV).

The problematic tests have been replaced with a TODO. Ideally, if and
when the MSRV reaches 1.10, it will be a simple matter of reverting this
commit to bring back the tests.