`TotalOrder` trait for floating point numbers
Define an orthogonal trait which corresponds to the `totalOrder`
predicate the IEEE 754 (2008 revision) floating point standard.

In order to maintain coherence, the bounds on `TotalOrder` should most
likely be `TotalOrder: Float` (or `TotalOrder: FloatCore`).  Without
type constraints, `TotalOrder` could be defined on, well, anything.
Though slightly ugly, one way to deal with this is to define two
traits, `TotalOrderCore: FloatCore` and `TotalOrder: Float`.  On the
other hand, `Inv` has no such constraints (due to the possibility of a
meaningful implementation on rational numbers).